chore: split Input and MultiCombo tests
