+++ packages/main/test/specs/input/MultiComboBox.spec.js
@@ -494,899 +494,6 @@ describe("MultiComboBox general interaction", () => {
 		});
 	});
 
-	describe("keyboard handling", () => {
-		before(async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-		});
-
-		it("tests backspace when combobox has an empty value", async () => {
-			let tokens = await browser.$("#multi1").shadow$$(".ui5-multi-combobox-token");
-			const input = await browser.$("#multi1").shadow$("input");
-
-			await input.click();
-			await input.keys('Backspace');
-
-			assert.strictEqual(tokens.length, 3, "3 tokens are visible");
-
-			await input.keys('Backspace');
-
-			tokens = await browser.$("#multi1").shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(tokens.length, 2, "2 tokens are visible");
-		});
-
-		it ("Value should be reset on ESC key", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mCombo = await browser.$("#another-mcb");
-			const mCombo2 = await browser.$("#more-mcb");
-			const input = await mCombo.shadow$("#ui5-multi-combobox-input");
-			const input2 = await mCombo2.shadow$("#ui5-multi-combobox-input");
-
-			await input.click();
-			await input.keys("C");
-			await input.keys("Escape");
-			await input.keys("Escape");
-
-			assert.strictEqual(await mCombo.getProperty("value"), "", "Value should be reset to the initial one");
-
-			await input.click();
-			await input.keys("C");
-
-			// Move focus to another element and bring it back
-			await input2.click();
-			await input.click();
-
-			await input.keys("o");
-			await input.keys("Escape");
-			await input.keys("Escape");
-
-			assert.strictEqual(await mCombo.getProperty("value"), "Cosy", "Value should be reset to the initial one");
-
-			await input2.click();
-			await input2.keys("C");
-			await input2.keys("Escape");
-
-			assert.strictEqual(await mCombo2.getProperty("value"), "", "Value should be cleared on escape even if the suggesitons are openjed");
-		});
-
-		it ("selects an item when enter is pressed and value matches a text of an item in the list", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-with-placeholder");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await input.keys(['c', 'o', 'm', 'p', 'a', 'c', 't']);
-			await input.keys("Enter");
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(tokens.length, 4, "4 tokens are visible");
-			assert.strictEqual(await input.getValue(), "", "Input's value should be empty");
-
-			await input.keys(['c', 'o', 's', 'y']);
-			await input.keys("Enter");
-
-			assert.strictEqual(await input.getValue(), "cosy", "value should remain cosy");
-			assert.strictEqual(await input.getAttribute("value-state"), "Error", "Value state is changed to error");
-			assert.strictEqual(await mcb.getProperty("valueStateDefaultText"), "This value is already selected.", "Value state text should be set to already selected");
-
-			await browser.waitUntil(async() => {
-				return await input.getAttribute("value-state") === "None";
-			}, 2500, "expect value state to be different after 2.5 seconds");
-		});
-
-		it ("focuses the value state header and item on arrow down then the value state and the input on arrow up", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-			const icon = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-error");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			const staticArea = await browser.execute(staticAreaItemClassName => document.querySelector(`.${staticAreaItemClassName}`), staticAreaItemClassName);
-
-			await icon.click();
-			await input.keys("ArrowDown");
-
-			let valueStateHeader = await browser.execute(staticArea => staticArea.shadowRoot.querySelector(".ui5-responsive-popover-header.ui5-valuestatemessage-root"), staticArea);
-			let focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header should be focused");
-
-			await input.keys("ArrowDown");
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[0];
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(await listItem.getProperty("focused"), true, "The first item is focused");
-			assert.notEqual(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header should not be focused");
-
-			await input.keys("ArrowUp");
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(await listItem.getProperty("focused"), false, "The first item is no longer focused");
-			assert.equal(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header is focused again");
-
-			await input.keys("ArrowUp");
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), true, "The input should be focused");
-			assert.equal(await listItem.getProperty("focused"), false, "The first item should not be focused");
-			assert.equal(focusedElement, null, "The value state header or item should not be focused");
-		});
-
-		it ("focuses the first item on arrow down, then the input on arrow up", async () => {
-			const mcb = await browser.$("#mcb-with-placeholder");
-			const input = await mcb.shadow$("input");
-			const icon = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-with-placeholder");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			const staticArea = await browser.execute(staticAreaItemClassName => document.querySelector(`.${staticAreaItemClassName}`), staticAreaItemClassName);
-
-			await icon.click();
-			await mcb.keys("ArrowDown");
-
-			const listItem = await popover.$("ui5-list").$$("ui5-li")[0];
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(await listItem.getProperty("focused"), true, "The first item is focused");
-
-			await input.keys("ArrowUp");
-
-			assert.equal(await mcb.getProperty("focused"), true, "The input should be focused");
-			assert.equal(await listItem.getProperty("focused"), false, "The first item is not focused");
-		});
-
-		it ("should navigate through the items with arrow keys when the picker is closed", async () => {
-			const mcb = await browser.$("#mcb");
-
-			await mcb.click();
-			await mcb.keys("ArrowDown");
-
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The first item name is seleted");
-
-			await mcb.keys("ArrowDown");
-			assert.equal(await mcb.getProperty("value"), "Compact", "The second item name is selected");
-
-			await mcb.keys("ArrowUp");
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The value is set back to the first item name");
-
-			await mcb.keys("ArrowUp");
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The value remains the same when pressing arrow up while the first item is set");
-		});
-
-		it ("should only navigate through not already selected items with arrow keys when the picker is closed", async () => {
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await input.keys("ArrowDown");
-
-			assert.equal(await mcb.getProperty("value"), "Compact", "The only not selected item is set");
-
-			await input.keys("ArrowDown");
-			assert.equal(await mcb.getProperty("value"), "Compact", "The only not selected item remains set");
-
-			await input.keys("ArrowUp");
-			assert.equal(await mcb.getProperty("value"), "Compact", "The only not selected item remains set");
-		});
-
-		it ("should reset current navigation state on user input", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-
-			await mcb.click();
-			await mcb.keys("ArrowDown");
-			await mcb.keys("ArrowDown");
-
-			assert.equal(await mcb.getProperty("value"), "Compact", "The second item name is selected");
-
-			await mcb.keys("a");
-			await mcb.keys("ArrowDown");
-
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The value is set to the first item name");
-
-			await mcb.keys("ArrowDown");
-			await mcb.keys("ArrowDown");
-
-			assert.equal(await mcb.getProperty("value"), "Condensed", "The value is set to the third item name");
-
-			await mcb.keys("a");
-			await mcb.keys("ArrowUp");
-			assert.equal(await mcb.getProperty("value"), "Longest word in the world 2", "The last item is selected if there is a custom value in the input");
-		});
-
-		it ("arrow up when no item is selected should go to the last item", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-
-			await mcb.click();
-			await mcb.keys("ArrowUp");
-
-			assert.equal(await mcb.getProperty("value"), "Longest word in the world 2", "Last value should be selected");
-		});
-
-		it ("Clicking delete icon should delete token and place the focus on the previous one", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			const secondToken = tokens[1];
-			const deleteIcon =  await tokens[0].shadow$("ui5-icon");
-
-			await secondToken.click();
-			await deleteIcon.click();
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 2, "should have two tokens");
-			assert.equal(await tokens[0].getProperty("focused"), true, "Previous token is focused");
-			assert.equal(await tokens[0].getProperty("text"), "Condensed", "The selected token should not be deleted.");
-		});
-
-		it ("BACKSPACE should delete token and place the focus on the previous one", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const input = await mcb.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			const secondToken = tokens[1];
-
-			await secondToken.click();
-			await mcb.keys("Backspace");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 2, "should have two tokens");
-			assert.equal(await tokens[0].getProperty("focused"), true, "Previous token is focused");
-		});
-
-		it ("DELETE should delete token and place the focus on the next one", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys("ArrowLeft");
-			await mcb.keys("ArrowLeft");
-
-			await mcb.keys("Delete");
-
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 2, "should have two tokens");
-			assert.equal(await tokens[1].getProperty("focused"), true, "Previous token is focused");
-		});
-
-		it ("BACKSPACE should delete token all selected tokens and place the focus on the first token before the deleted ones", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-items");
-			const input = await mcb.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			const fourthToken = tokens[3];
-
-			await fourthToken.click();
-			await mcb.keys(["Shift", "ArrowLeft"]);
-			await mcb.keys("Backspace");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			assert.equal(await tokens.length, 3, "should have three tokens");
-			assert.equal(await tokens[1].getProperty("focused"), true, "Second token is focused");
-		});
-
-		it ("DELETE should delete token all selected tokens and place the focus on the first token after the deleted ones", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-items");
-			const input = await mcb.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			const fourthToken = tokens[3];
-
-			await fourthToken.click();
-			await mcb.keys(["Shift", "ArrowLeft"]);
-			await mcb.keys("Delete");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 3, "should have two tokens");
-			assert.equal(await tokens[2].getProperty("focused"), true, "Last token is focused");
-		});
-
-		it ("should focus input after all tokens are deleted", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-compact");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys("ArrowLeft");
-			await mcb.keys(["Shift", "ArrowLeft"]);
-			await mcb.keys("Delete");
-
-			assert.equal(await mcb.getProperty("focused"), true, "Input should be focused");
-		});
-
-		it ("first HOME should move caret to start of the input, second HOME should focus the first token, END should focus last token", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await mcb.click();
-			await input.keys(['c', 'o', 'm', 'p']);
-			await input.keys("Home");
-
-			const caretPosition = await browser.executeAsync(done => {
-				return done(document.querySelector("#mcb").shadowRoot.querySelector("input").selectionStart);
-			});
-
-			assert.equal(await caretPosition, 0, "The caret is at the start of the input");
-
-			await input.keys("Home");
-			assert.equal(await tokens[0].getProperty("focused"), true, "The first token is focused");
-
-			await input.keys("End");
-			assert.equal(await tokens[tokens.length - 1].getProperty("focused"), true, "The last token is focused");
-
-			await input.keys("End");
-			assert.equal(await mcb.getProperty("focused"), true, "The input is focused");
-		});
-
-		it ("CTRL + HOME focus the first token, CTRL + END should focus last token", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await input.keys(['Control','Home']);
-
-			assert.equal(await tokens[0].getProperty("focused"), true, "The first token is focused");
-
-			await input.keys(['Control','End']);
-			assert.equal(await tokens[tokens.length - 1].getProperty("focused"), true, "The last token is focused");
-		});
-
-		it ("CTRL + HOME focus the first item, CTRL + END should focus last item", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const input = await mcb.shadow$("input");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			const staticArea = await browser.execute(staticAreaItemClassName => document.querySelector(`.${staticAreaItemClassName}`), staticAreaItemClassName);
-
-
-			await input.click();
-			await mcb.keys("F4");
-			await mcb.keys("ArrowDown");
-			await mcb.keys(['Control','End']);
-
-			const lastItem = await popover.$("ui5-list").$$("ui5-li")[5];
-			assert.equal(await lastItem.getProperty("focused"), true, "The last item is focused");
-
-			await mcb.keys(['Control','Home']);
-
-			const firstItem = await popover.$("ui5-list").$("ui5-li");
-			assert.equal(await firstItem.getProperty("focused"), true, "The first item is focused");
-
-		});
-
-		it ("SHIFT + HOME should select all tokens from the current one to the first one", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const input = await mcb.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await input.keys(['Shift','Home']);
-
-			assert.equal(await tokens[0].getProperty("focused"), true, "The first token is focused");
-			assert.equal(await tokens[0].getProperty("selected"), true, "The first token is selected");
-			assert.equal(await tokens[1].getProperty("selected"), true, "The second token is selected");
-			assert.equal(await tokens[2].getProperty("selected"), false, "The last token is not selected");
-		});
-
-		it ("SHIFT + END should select all tokens from the current one to the last one", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const input = await mcb.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await input.keys(['Shift','End']);
-
-			assert.equal(await tokens[2].getProperty("focused"), true, "The last token is focused");
-			assert.equal(await tokens[2].getProperty("selected"), true, "The last token is selected");
-			assert.equal(await tokens[1].getProperty("selected"), true, "The second token is selected");
-			assert.equal(await tokens[0].getProperty("selected"), false, "The first token is not selected");
-		});
-
-		it ("should close the picker and focus the next element on TAB", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const mcb2 = await browser.$("#mcb-no-typeahead");
-
-			await mcb.click();
-			await mcb.keys("F4");
-			await mcb.keys("Tab");
-
-			assert.equal(await mcb.getProperty("open"), false, "The previous control is closed");
-			assert.equal(await mcb2.getProperty("focused"), true, "The next control is focused");
-
-			await mcb2.keys("F4");
-			await mcb2.keys(["Shift", "Tab"]);
-
-			assert.equal(await mcb2.getProperty("open"), false, "The next control is closed");
-			assert.equal(await mcb.getProperty("focused"), true, "The previous control is focused");
-
-			await mcb.keys(["a", "b", "c"]);
-			await mcb.keys("F4");
-			await mcb.keys("Tab");
-
-			assert.equal(await mcb.getProperty("open"), false, "The previous control is closed");
-			assert.equal(await mcb.getProperty("value"), "abc", "The previous control value is set");
-			assert.equal(await mcb2.getProperty("focused"), true, "The next control is focused");
-		});
-
-		it ("should close the picker and focus the next element on TAB over an item or value state header", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-warning");
-			const input = await mcb.shadow$("input");
-			const mcb2 = await browser.$("#mcb-error");
-
-			await input.click();
-			await mcb.keys("F4");
-			await mcb.keys("ArrowDown");
-			await mcb.keys("Tab");
-
-			assert.equal(await mcb.getProperty("open"), false, "The previous control is closed after TAB on value state header");
-			assert.equal(await mcb2.getProperty("focused"), true, "The next control is focused after TAB on value state header");
-
-			await mcb2.keys("F4");
-			await mcb2.keys("ArrowDown");
-			await mcb2.keys("ArrowDown");
-			await mcb2.keys(["Shift", "Tab"]);
-
-			assert.equal(await mcb2.getProperty("open"), false, "The next control is closed after TAB on suggestion item");
-			assert.equal(await mcb.getProperty("focused"), true, "The previous control is focused after TAB on suggestion item");
-
-			await mcb.keys("F4");
-			await mcb.keys("ArrowDown");
-			await mcb.keys("ArrowDown");
-			await mcb.keys("Tab");
-
-			assert.equal(await mcb.getProperty("open"), false, "The previous control is closed after TAB on suggestion item");
-			assert.equal(await mcb2.getProperty("focused"), true, "The next control is focused after TAB on suggestion item");
-		});
-
-		it ("should select/unselect next/previous item on shift+arrow", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-
-			await mcb.click();
-			await mcb.keys("F4");
-			await mcb.keys("ArrowDown");
-			await mcb.keys("Space");
-			await mcb.keys(["Shift", "ArrowDown"]);
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(tokens.length, 2, "should have two items selected");
-
-			await mcb.keys("ArrowDown");
-			await mcb.keys(["Shift", "ArrowUp"]);
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(tokens.length, 1, "should have two items selected");
-		});
-
-		it ("should move focus to the previous token with arrow left", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys("ArrowLeft");
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(await tokens[2].getProperty("focused"), true, "Last token should be focused");
-
-			await mcb.keys("ArrowLeft");
-			assert.strictEqual(await tokens[1].getProperty("focused"), true, "Second token should be focused");
-
-			await mcb.keys("ArrowLeft");
-			assert.strictEqual(await tokens[0].getProperty("focused"), true, "First token should be focused");
-		});
-
-		it ("should select multiple tokens and move focus with shift+arrow keys", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const mcb2 = await browser.$("#mcb-warning");
-			const input = await mcb.shadow$("input");
-			const input2 = await mcb2.shadow$("input");
-
-
-			await input.click();
-			await mcb.keys("ArrowLeft");
-			await mcb.keys(["Shift", "ArrowLeft"]);
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(await tokens[2].getProperty("selected"), true, "Last token should be selected");
-			assert.strictEqual(await tokens[1].getProperty("selected"), true, "Second token should be selected");
-			assert.strictEqual(await tokens[1].getProperty("focused"), true, "Second token should be focused");
-
-			await mcb.keys(["Shift", "ArrowLeft"]);
-
-			assert.strictEqual(await tokens[2].getProperty("selected"), true, "Last token should be selected");
-			assert.strictEqual(await tokens[1].getProperty("selected"), true, "Second token should be selected");
-			assert.strictEqual(await tokens[0].getProperty("selected"), true, "First token should be selected");
-			assert.strictEqual(await tokens[0].getProperty("focused"), true, "First token should be focused");
-
-			await input2.click();
-			await mcb2.keys("ArrowLeft");
-			await mcb2.keys("ArrowLeft");
-			await mcb2.keys("ArrowLeft");
-
-			await mcb2.keys(["Shift", "ArrowRight"]);
-
-			let tokens2 = await mcb2.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(await tokens2[0].getProperty("selected"), true, "First token should be selected");
-			assert.strictEqual(await tokens2[1].getProperty("selected"), true, "Second token should be selected");
-			assert.strictEqual(await tokens2[1].getProperty("focused"), true, "second token should be focused");
-
-			await mcb2.keys(["Shift", "ArrowRight"]);
-
-			assert.strictEqual(await tokens2[2].getProperty("focused"), true, "Last token should be focused");
-			assert.strictEqual(await tokens2[2].getProperty("selected"), true, "Last token should be selected");
-			assert.strictEqual(await tokens2[1].getProperty("selected"), true, "Second token should be selected");
-			assert.strictEqual(await tokens2[0].getProperty("selected"), true, "First token should be selected");
-		});
-
-		it ("should navigate through the items with CTRL + arrow up/down keys when the picker is closed", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-
-			await mcb.click();
-			await mcb.keys(["Control", "ArrowDown"]);
-
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The first item name is selected");
-
-			await mcb.keys(["Control", "ArrowDown"]);
-			assert.equal(await mcb.getProperty("value"), "Compact", "The second item name is selected");
-
-			await mcb.keys(["Control", "ArrowUp"]);
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The value is set back to the first item name");
-
-			await mcb.keys(["Control", "ArrowUp"]);
-			assert.equal(await mcb.getProperty("value"), "Cosy", "The value remains the same when pressing arrow up while the first item is set");
-		});
-
-		it ("focuses the value state header and item on CTRL + arrow down then the value state and the input on CTRL + arrow up", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-			const icon = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-error");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			const staticArea = await browser.execute(staticAreaItemClassName => document.querySelector(`.${staticAreaItemClassName}`), staticAreaItemClassName);
-
-			await icon.click();
-			await mcb.keys(["Control", "ArrowDown"]);
-
-			let valueStateHeader = await browser.execute(staticArea => staticArea.shadowRoot.querySelector(".ui5-responsive-popover-header.ui5-valuestatemessage-root"), staticArea);
-			let focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header should be focused");
-
-			await mcb.keys(["Control", "ArrowDown"]);
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[0];
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(await listItem.getProperty("focused"), true, "The first item is focused");
-			assert.notEqual(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header should not be focused");
-
-			await mcb.keys(["Control", "ArrowUp"]);
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), false, "The input should not be focused");
-			assert.equal(await listItem.getProperty("focused"), false, "The first item is no longer focused");
-			assert.equal(focusedElement[Object.keys(focusedElement)[0]], valueStateHeader[Object.keys(valueStateHeader)[0]], "The value state header is focused again");
-
-			await mcb.keys(["Control", "ArrowUp"]);
-			focusedElement = await browser.execute(staticArea => staticArea.shadowRoot.activeElement, staticArea);
-
-			assert.equal(await mcb.getProperty("focused"), true, "The input should be focused");
-			assert.equal(await listItem.getProperty("focused"), false, "The first item should not be focused");
-			assert.equal(focusedElement, null, "The value state header or item should not be focused");
-		});
-
-		it ("should select all filtered items on CTRL+A", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys("F4");
-			await mcb.keys("ArrowDown");
-			await mcb.keys(["Control", "a"]);
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 6, "All items are selected");
-
-			await mcb.keys(["Control", "a"]);
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 0, "All items are deselected");
-
-			await input.click();
-			await mcb.keys("c");
-			await mcb.keys("ArrowDown");
-			await mcb.keys(["Control", "a"]);
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 3, "All filtered items are selected");
-
-			await mcb.keys(["Control", "a"]);
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.equal(await tokens.length, 0, "All selected filtered items are deselected");
-		});
-
-		it ("should copy a token with CTRL+C and paste it with CTRL+V", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const mcb2 = await browser.$("#mcb");
-			const input = await mcb2.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await tokens[1].keys(["Control", "c"]);
-			await input.click();
-			await input.keys(["Control", "v"]);
-
-			assert.equal(await mcb2.getProperty("value"), "Condensed", "Token is pasted into the second control");
-		});
-
-		it ("should not be able to paste tokenwith CTRL+V in read only multi combo box", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const mcb2 = await browser.$("#readonly-value-state-mcb");
-			const input = await mcb2.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await tokens[1].keys(["Control", "c"]);
-			await input.click();
-			await input.keys(["Control", "v"]);
-
-			const mcb2Tokens = await mcb2.shadow$$(".ui5-multi-combobox-token");
-			assert.equal(await mcb2.getProperty("value"), "", "Token is not pasted into the second control");
-			assert.equal(mcb2Tokens.length, 0, "No token was created.");
-		});
-
-		it ("should cut a token with CTRL+X and paste it with CTRL+V", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const mcb2 = await browser.$("#mcb");
-			const input = await mcb2.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await tokens[1].keys(["Control", "x"]);
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			assert.equal(await tokens.length, 2, "One of the tokens is cut from the control");
-
-			await input.click();
-			await input.keys(["Control", "v"]);
-
-			assert.equal(await mcb2.getProperty("value"), "Condensed", "Token is pasted into the second control");
-		});
-
-		it ("should cut a token with SHIFT+DELETE and paste it with SHIFT+INSERT", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const mcb2 = await browser.$("#mcb");
-			const input = await mcb2.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await tokens[1].keys(["Shift", "Delete"]);
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			assert.equal(await tokens.length, 2, "One of the tokens is cut from the control");
-
-			await input.click();
-			await input.keys(["Shift", "Insert"]);
-
-			assert.equal(await mcb2.getProperty("value"), "Condensed", "Token is pasted into the second control");
-		});
-
-		it ("should copy a token with CTRL+INSERT and paste it with SHIFT+INSERT", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi1");
-			const mcb2 = await browser.$("#mcb");
-			const input = await mcb2.shadow$("input");
-			const tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[1].click();
-			await tokens[1].keys(["Control", "Insert"]);
-			await input.click();
-			await input.keys(["Shift", "Insert"]);
-
-			assert.equal(await mcb2.getProperty("value"), "Condensed", "Token is pasted into the second control");
-		});
-
-		it ("should select a token with CTRL+SPACE", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-error");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys("ArrowLeft");
-			await mcb.keys(["Control", "Space"]);
-
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(await tokens[2].getProperty("selected"), true, "Last token should be selected");
-		});
-
-		it ("CTRL+SPACE should do nothing when pressed in the input field", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const input = await mcb.shadow$("input");
-
-			await input.click();
-			await mcb.keys(["Shift", "Space"]);
-
-			assert.strictEqual(await mcb.getProperty("value"), "", "Input field is empty");
-		});
-
-		it ("F4 should focus the selected item or the first one if there is no selected", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-
-			await mcb.click();
-			await mcb.keys("F4");
-
-			const listItem = await popover.$("ui5-list").$("ui5-li");
-
-			assert.equal(await listItem.getProperty("focused"), true, "The first item is focused");
-
-			await mcb.keys("ArrowDown");
-			await mcb.keys("Space");
-			await mcb.keys("F4");
-			await mcb.keys("F4");
-
-			const listItem2 = await popover.$("ui5-list").$$("ui5-li")[1];
-
-			assert.equal(await listItem2.getProperty("focused"), true, "The second item is focused as it is selected");
-		});
-
-		it ("Alt + Down should focus the corresponding item to the token from which the combination is pressed", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-items");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-items");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await tokens[2].click();
-			await tokens[2].keys(["Alt", "ArrowDown"]);
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[3];
-
-			assert.equal(await listItem.getProperty("focused"), true, "The selected item corresponding to the token is focused");
-		});
-
-		it ("Alt + Down should focus the first item if no selected items are present", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#multi-acv");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#multi-acv");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-
-			await mcb.click();
-			await mcb.keys(["Alt", "ArrowDown"]);
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[0];
-			assert.equal(await listItem.getProperty("focused"), true, "The first item is focused");
-		});
-
-		it ("Alt + Down should not filter items", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const input = await mcb.shadow$("input");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-
-			await input.click();
-			await input.keys(["a", "b"]);
-			await input.keys(["Alt", "ArrowDown"]);
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[0];
-
-			assert.equal(await listItem.getProperty("focused"), true, "The items are not filtered and the first item is focused");
-		});
-
-		it ("Alt + Down should focus the item corresponding to the text value", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb");
-			const input = await mcb.shadow$("input");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-
-			await input.click();
-			await input.keys("ArrowDown");
-			await input.keys("ArrowDown");
-			await input.keys(["Alt", "ArrowDown"]);
-
-			let listItem = await popover.$("ui5-list").$$("ui5-li")[1];
-
-			assert.equal(await listItem.getProperty("focused"), true, "The second item should be focused");
-		});
-
-		it ("Backspace deletes token and forwards the focus to the last token without collapsing the tokenizer", async () => {
-			const mcb = await $("#n-more-many-items");
-			const inner = await mcb.shadow$("input");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			await inner.click();
-
-			assert.strictEqual(tokens.length, 7, "7 Tokens are placed in the MCB");
-
-			await inner.keys("Backspace");
-			await inner.keys("Backspace");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-
-			assert.strictEqual(tokens.length, 6, "6 Tokens are placed in the MCB");
-			assert.ok(await tokens[tokens.length - 1].getProperty("focused"), "Last Token is focused");
-		});
-	});
-
 	describe("General", () => {
 		before(async () => {
 			await browser.url(`test/pages/MultiComboBox.html`);
@@ -1473,412 +580,6 @@ describe("MultiComboBox general interaction", () => {
 
 			await mcb.click();
 			assert.notOk(await popover.getProperty("opened"), "Popover with valueStateMessage should not be opened.");
-		});
-	});
-
-	describe("ARIA attributes", () => {
-		before(async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-		});
-
-		it ("aria-describedby value according to the tokens count and the value state", async () => {
-			const mcb = await browser.$("#mcb-error");
-			const innerInput = await mcb.shadow$("input");
-			const invisibleText = await mcb.shadow$(".ui5-hidden-text");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			const tokensCountITextId = `ui5-multi-combobox-hiddenText-nMore`;
-			const valuestateITextId = `ui5-multi-combobox-valueStateDesc`;
-			const ariaDescribedBy = `${tokensCountITextId} ${valuestateITextId}`;
-
-			assert.strictEqual(tokens.length, 3, "should have three tokens");
-			assert.strictEqual(await innerInput.getAttribute("aria-describedby"), ariaDescribedBy, "aria-describedby has a reference for the value state and the tokens count");
-		});
-
-		it ("aria-describedby value according to the tokens count", async () => {
-			const mcb = await browser.$("#mcb-compact");
-
-			await mcb.scrollIntoView();
-
-			const innerInput = await mcb.shadow$("input");
-			let invisibleText = await mcb.shadow$(".ui5-hidden-text");
-			const inivisbleTextId = await invisibleText.getProperty("id");
-			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			let resourceBundleText = null;
-
-			assert.strictEqual(tokens.length, 2, "should have two tokens");
-			assert.strictEqual(await innerInput.getAttribute("aria-describedby"), inivisbleTextId, "aria-describedby reference is correct");
-			// assert.strictEqual(await invisibleText.getText(), "Contains 2 tokens", "aria-describedby text is correct");
-
-			await innerInput.click();
-			await innerInput.keys("Backspace");
-			await innerInput.keys("Backspace");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			invisibleText = await mcb.shadow$(".ui5-hidden-text");
-
-			resourceBundleText = await browser.executeAsync(done => {
-				const mcb = document.getElementById("mcb-compact");
-				done(mcb.constructor.i18nBundle.getText(window["sap-ui-webcomponents-bundle"].defaultTexts.TOKENIZER_ARIA_CONTAIN_ONE_TOKEN));
-			});
-
-			assert.strictEqual(tokens.length, 1, "should have one token");
-			// assert.strictEqual(await invisibleText.getText(), resourceBundleText, "aria-describedby text is correct");
-
-			await innerInput.keys("Backspace");
-
-			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
-			invisibleText = await mcb.shadow$(".ui5-hidden-text");
-
-			resourceBundleText = await browser.executeAsync(done => {
-				const mcb = document.getElementById("mcb-compact");
-				done(mcb.constructor.i18nBundle.getText(window["sap-ui-webcomponents-bundle"].defaultTexts.TOKENIZER_ARIA_CONTAIN_TOKEN));
-			});
-
-			assert.strictEqual(tokens.length, 0, "should not have tokens");
-			// assert.strictEqual(await invisibleText.getText(), resourceBundleText, "aria-describedby text is correct");
-		});
-
-		it ("Should apply aria-label from the accessibleName property", async () => {
-			const mcb = await browser.$("#multi1");
-			const innerInput = await mcb.shadow$("input");
-
-			await mcb.scrollIntoView();
-
-			assert.strictEqual(await innerInput.getAttribute("aria-label"), "MultiComboBox with predefined values", "aria-label attribute is correct.");
-		});
-
-		it ("Should apply aria-label from the accessibleNameRef property", async () => {
-			const mcb = await browser.$("#mcb-predefined-value");
-			const innerInput = await mcb.shadow$("input");
-			const mcbLabel = await browser.$("#mcbLabel");
-
-			await mcb.scrollIntoView();
-
-			assert.strictEqual(await innerInput.getAttribute("aria-label"), await mcbLabel.getHTML(false), "aria-label attribute is correct.");
-		});
-
-		it("Value state type should be added to the screen readers default value states announcement", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mCbWarning = await browser.$("#mcb-warning");
-			const mCbSuccess = await browser.$("#mcb-success");
-			const mCbError = await browser.$("#mcb-error");
-
-			let staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-warning");
-			let popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-popover");
-
-			await mCbWarning.click();
-
-			let ariaHiddenText = await mCbWarning.shadow$(`#ui5-multi-combobox-valueStateDesc`).getHTML(false);
-			let valueStateText = await popover.$("div").getHTML(false);
-
-			assert.strictEqual(ariaHiddenText.includes("Value State"), true, "Hidden screen reader text is correct");
-			assert.strictEqual(valueStateText.includes("Warning issued"), true, "Displayed value state message text is correct");
-
-			await mCbWarning.keys("Escape");
-			await mCbError.click();
-
-			staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-error");
-			popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-popover");
-
-			ariaHiddenText = await mCbError.shadow$(`#ui5-multi-combobox-valueStateDesc`).getHTML(false);
-			valueStateText = await popover.$("div").getHTML(false);
-
-			assert.strictEqual(ariaHiddenText.includes("Value State"), true, "Hidden screen reader text is correct");
-			assert.strictEqual(valueStateText.includes("Invalid entry"), true, "Displayed value state message text is correct");
-
-			await mCbError.keys("Escape");
-			await mCbSuccess.click();
-
-			staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-success");
-			ariaHiddenText = await mCbSuccess.shadow$(`#ui5-multi-combobox-valueStateDesc`).getHTML(false);
-
-			assert.strictEqual(ariaHiddenText.includes("Value State"), true, "Hidden screen reader text is correct");
-		});
-
-		it("Value state type should be added to the screen readers custom value states announcement", async () => {
-			const mCbInformation = await browser.$("#mcb-information");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-information");
-
-			await mCbInformation.click();
-			await mCbInformation.keys("a");
-
-			const popoverHeader = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover .ui5-valuestatemessage-header");
-			const valueStateText = await popoverHeader.$("div").getHTML(false);
-			const ariaHiddenText = await mCbInformation.shadow$(`#ui5-multi-combobox-valueStateDesc`).getHTML(false);
-
-			assert.strictEqual(ariaHiddenText.includes("Value State"), true, "Hidden screen reader text is correct");
-			assert.strictEqual(valueStateText.includes("Extra long text used as an information message"), true, "Displayed value state message text is correct");
-		});
-	});
-
-	describe("Grouping", () => {
-		it ("Tests group filtering", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			let popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			let listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			await arrow.click();
-
-			assert.strictEqual(groupItems.length, 3, "Group items should be 3");
-			assert.strictEqual(listItems.length, 12, "Items should be 12");
-
-			await input.keys("B");
-
-			popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			assert.strictEqual(groupItems.length, 1, "Filtered group items should be 1");
-			assert.strictEqual(listItems.length, 1, "Filtered items should be 1");
-
-			await input.keys("Backspace");
-			await input.keys(['E', 'u', 'r', 'o', 'p', 'e']);
-
-			assert.equal(await popover.getProperty("opened"), false, "Popover should not be open");
-		});
-
-		it ("Tests group item focusability", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-		});
-
-		it ("Group header keyboard handling", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-			await groupItem.keys("Enter");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated");
-
-			await groupItem.keys("Space");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated)");
-
-			await groupItem.keys("ArrowUp");
-
-			assert.equal(await groupItem.getProperty("focused"), false, "The first group header should be focused");
-			assert.equal(await mcb.getProperty("focused"), true, "The first group header should be focused");
-		});
-
-		it ("Should not select group headers", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-
-			await input.click();
-			await mcb.setProperty("value", "Asia");
-			await mcb.keys("Enter");
-
-			let tokens = await browser.$("#mcb-grouping").shadow$$(".ui5-multi-combobox-token");
-			let selectionChangeFired = await browser.execute(() => document.getElementById("selection-change-events-fired").textContent);
-
-			assert.strictEqual(await input.getValue(), "Asia", "The value remains");
-			assert.strictEqual(tokens.length, 0, "The group header is not tokenized");
-			assert.strictEqual(selectionChangeFired, "", "SelectionChange event is not fired");
-		});
-	});
-
-	describe("Grouping", () => {
-		it ("Tests group filtering", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			let popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			let listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			await arrow.click();
-
-			assert.strictEqual(groupItems.length, 3, "Group items should be 3");
-			assert.strictEqual(listItems.length, 12, "Items should be 12");
-
-			await input.keys("B");
-
-			popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			assert.strictEqual(groupItems.length, 1, "Filtered group items should be 1");
-			assert.strictEqual(listItems.length, 1, "Filtered items should be 1");
-
-			await input.keys("Backspace");
-			await input.keys(['E', 'u', 'r', 'o', 'p', 'e']);
-
-			assert.equal(await popover.getProperty("opened"), false, "Popover should not be open");
-		});
-
-		it ("Tests group item focusability", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-		});
-
-		it ("Group header keyboard handling", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-			await groupItem.keys("Enter");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated");
-
-			await groupItem.keys("Space");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated)");
-
-			await groupItem.keys("ArrowUp");
-
-			assert.equal(await groupItem.getProperty("focused"), false, "The first group header should be focused");
-			assert.equal(await mcb.getProperty("focused"), true, "The first group header should be focused");
-		});
-	});
-
-	describe("Grouping", () => {
-		it ("Tests group filtering", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			let popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			let listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			await arrow.click();
-
-			assert.strictEqual(groupItems.length, 3, "Group items should be 3");
-			assert.strictEqual(listItems.length, 12, "Items should be 12");
-
-			await input.keys("B");
-
-			popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			groupItems = await popover.$("ui5-list").$$("ui5-li-groupheader");
-			listItems = await popover.$("ui5-list").$$("ui5-li");
-
-			assert.strictEqual(groupItems.length, 1, "Filtered group items should be 1");
-			assert.strictEqual(listItems.length, 1, "Filtered items should be 1");
-
-			await input.keys("Backspace");
-			await input.keys(['E', 'u', 'r', 'o', 'p', 'e']);
-
-			assert.equal(await popover.getProperty("opened"), false, "Popover should not be open");
-		});
-
-		it ("Tests group item focusability", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-		});
-
-		it ("Group header keyboard handling", async () => {
-			await browser.url(`test/pages/MultiComboBox.html`);
-
-			const mcb = await browser.$("#mcb-grouping");
-			const input = await mcb.shadow$("#ui5-multi-combobox-input");
-			const arrow = await mcb.shadow$("[input-icon]");
-			const staticAreaItemClassName = await browser.getStaticAreaItemClassName("#mcb-grouping");
-			const popover = await browser.$(`.${staticAreaItemClassName}`).shadow$("ui5-responsive-popover");
-			let groupItem;
-
-			await arrow.click();
-			await input.keys("ArrowDown");
-
-
-			groupItem = await popover.$("ui5-list").$$("ui5-li-groupheader")[0];
-			await groupItem.keys("Enter");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated");
-
-			await groupItem.keys("Space");
-
-			assert.equal(await groupItem.getProperty("focused"), true, "The first group header should be focused");
-			assert.equal(await popover.getProperty("opened"), true, "Popover should not be open");
-			assert.strictEqual(await input.getValue(), "", "The value is not updated)");
-
-			await groupItem.keys("ArrowUp");
-
-			assert.equal(await groupItem.getProperty("focused"), false, "The first group header should be focused");
-			assert.equal(await mcb.getProperty("focused"), true, "The first group header should be focused");
 		});
 	});
 });
